requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.30.1
pydantic==2.10.1
pydantic-settings==2.10.1
python-dotenv==1.0.1
httpx==0.27.0
certifi==2025.10.5
SQLAlchemy==2.0.44
alembic==1.17.0
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
